requirements.txt
langchain==0.3.3
langchain_community==0.3.2
langchain_openai==0.2.2
faiss-cpu==1.9.0
langgraph==0.2.35
fastapi==0.115.2
sse_starlette==2.1.3
streamlit==1.39.0
langserve==0.3.0
zhipuai
httpx-sse==0.4.0
python-dotenv
uvicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib